requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
motor==3.6.0
pymongo==4.10.1
